replaced initializing of firebase into services->app.js
