--- ios-planner/src/services/firestoreMethods.js
@@ -1,17 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyB1CnrIALwWHoRfS4mJQrPDuOWsC90lR3Q",
-  authDomain: "ios-planner-709b1.firebaseapp.com",
-  projectId: "ios-planner-709b1",
-  storageBucket: "ios-planner-709b1.appspot.com",
-  messagingSenderId: "556738130951",
-  appId: "1:556738130951:web:3e1efb8edab2ab9b69e6a2"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
